+++ frontend/package.json
@@ -45,6 +45,7 @@
   "devDependencies": {
     "@rsbuild/core": "^0.6.4",
     "@rsbuild/plugin-react": "^0.6.4",
+    "@types/jest": "^29.5.12",
     "eslint": "^8.57.0",
     "eslint-plugin-react": "^7.34.1"
   },
