+++ frontend/src/App.css
@@ -1,6 +1,6 @@
-.App {
+/* .App {
   
-}
+} */
 
 /* .App-logo {
   height: 40vmin;
