+++ frontend/src/App.tsx
@@ -4,13 +4,17 @@ import { BrowserRouter } from "react-router-dom";
 
 import "./App.css";
 import AppRoutes from "./AppRoutes";
+import { ConfigProvider } from "antd";
+import { theme } from "./config/style";
 
 function App() {
   return (
     <div className="App">
-      <BrowserRouter>
-        <AppRoutes />
-      </BrowserRouter>
+      <ConfigProvider theme={theme}>
+        <BrowserRouter>
+          <AppRoutes />
+        </BrowserRouter>
+      </ConfigProvider>
     </div>
   );
 }
