+++ frontend/src/config/style.ts
@@ -1,0 +1,7 @@
+import { ThemeConfig } from "antd";
+
+export const theme: ThemeConfig = {
+  token: {
+    colorPrimary: "red",
+  },
+};
